fix: transaction value float field type to decimal

+++ src/api/schemas/balance_schemas.py
@@ -1,5 +1,7 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class BalanceResponse(BaseModel):
-    balance: float
+    balance: Decimal
